+++ ghost/core/core/server/services/adapter-manager/AdapterManager.js
